feat/refactor-components: extract DateInfo from MatchTile as a component
